chore(analytics-core): move core client under types
