+++ packages/analytics-core/src/timeline.ts
@@ -1,5 +1,5 @@
 import { AnalyticsIdentity, BeforePlugin, DestinationPlugin, EnrichmentPlugin, Plugin } from './types/plugin';
-import { CoreClient } from './core-client';
+import { CoreClient } from './types/client/core-client';
 import { IConfig } from './types/config/core-config';
 import { ILogger } from './logger';
 import { EventCallback } from './types/event-callback';
