+++ packages/analytics-core/src/types/browser-client.ts
@@ -1,5 +1,5 @@
 import { TransportType } from './transport';
-import { CoreClient } from '../core-client';
+import { CoreClient } from './client/core-client';
 import { AnalyticsIdentity, Plugin } from './plugin';
 import { AmplitudeReturn } from '../utils/return-wrapper';
 import { BrowserConfig, BrowserOptions } from './config/browser-config';
